OPS-34: refactor swipe up directions and map popup
